requirements.txt
affine==2.4.0
alembic==1.16.5
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.0.5
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
beautifulsoup4==4.12.2
bleach==4.1.0
bokeh==3.6.0
bottle==0.13.4
Bottleneck==1.3.7
Brotli==1.0.9
Cartopy==0.23.0
cdsapi==0.6.1
certifi==2024.8.30
cffi==1.16.0
cftime==1.6.2
charset-normalizer==2.0.4
choreographer==1.0.10
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
colorcet==3.1.0
colorlog==6.9.0
comm==0.2.1
contourpy==1.2.0
crc32c==2.7.1
cycler==0.12.1
cytoolz==0.12.2
dask==2024.8.2
dask-expr==1.1.13
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distributed==2024.8.2
donfig==0.8.1.post1
executing==0.8.3
fast-histogram==0.14
fasteners==0.19
fastjsonschema==2.16.2
filelock==3.13.1
fonttools==4.50.0
fqdn==1.5.1
fsspec==2024.6.1
greenlet==3.2.4
HeapDict==1.0.1
idna==3.7
importlib_metadata==8.5.0
intel-cmplr-lib-ur==2025.0.0
intel-openmp==2025.0.0
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.6
jsonpointer==3.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab_widgets==3.0.13
kaleido==1.0.0
kiwisolver==1.4.5
llvmlite==0.43.0
lmdb==1.4.1
locket==1.0.0
logistro==1.1.0
lz4==4.3.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mkl==2025.0.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.1
mpl-scatter-density==0.8
mpmath==1.3.0
msgpack==1.0.3
munkres==1.1.4
narwhals==2.3.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
netCDF4==1.6.2
networkx==3.2.1
notebook==7.0.8
notebook_shim==0.2.3
npy-append-array==0.9.16
numba==0.60.0
numcodecs==0.15.1
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
optuna==4.5.0
optuna-dashboard==0.19.0
orjson==3.11.3
overrides==7.4.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==3.10.0
plotly==6.3.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
properscoring==0.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.21
Pygments==2.15.1
pymannkendall==1.4.3
pyparsing==3.1.2
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.15
PyQt5-sip==12.13.0
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==25.1.2
rasterio==1.3.10
referencing==0.30.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rioxarray==0.15.5
rpds-py==0.10.6
scikit-learn==1.5.1
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
shapely==2.0.1
simplejson==3.20.1
sip==6.7.12
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.43
stack-data==0.2.0
statsmodels==0.14.2
sympy==1.13.1
tbb==2022.0.0
tblib==1.7.0
tcmlib==1.2.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.7.1
triton==3.1.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.15.0
tzdata==2023.3
umf==0.9.0
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==0.58.0
widgetsnbextension==4.0.13
win-inet-pton==1.1.0
wrapt==1.17.2
xarray==2025.6.1
xhistogram==0.3.2
xskillscore==0.0.26
xyzservices==2022.9.0
zarr==3.0.2
zict==3.0.0
zipp==3.21.0
